requirements.txt
torch
torchvision
opencv-python
numpy
pillow
ultralytics
transformers
scikit-image
